feat: place duplicate data functionality in data utilities and add duplication to the message system
